♻️ Extract public/rows.go from rows.go
