+++ public/rows.go
@@ -1,0 +1,21 @@
+package public
+
+import (
+	"errors"
+	"io"
+	"log/slog"
+)
+
+// HandleRowsNextError returns a boolean indicating completion and a slice of slog.Attr.
+// If err is nil, it returns true and a slice of slog.Attr{slog.Bool("eof", false)}.
+// If err is io.EOF, it returns true and a slice of slog.Attr{slog.Bool("eof", true)}.
+// Otherwise, it returns false and nil.
+func HandleRowsNextError(err error) (bool, []slog.Attr) {
+	if err == nil {
+		return true, []slog.Attr{slog.Bool("eof", false)}
+	}
+	if errors.Is(err, io.EOF) {
+		return true, []slog.Attr{slog.Bool("eof", true)}
+	}
+	return false, nil
+}
